+++ db/migrate/20201019004448_create_comments.rb
@@ -1,8 +1,8 @@
 class CreateComments < ActiveRecord::Migration[5.2]
   def change
     create_table :comments do |t|
-      t.integer "user_id", null: false
-      t.integer "review_id", null: false
+      t.bigint "user_id", null: false
+      t.bigint "review_id", null: false
       t.string "title", null: false
       t.text "body", null: false
       t.timestamps
